requirements.txt
yt-dlp
opencv-python
tqdm
label-studio
label-studio-ml
ultralytics
supervision
label-studio-sdk
numpy
filetype
loguru
pyyaml
scipy==1.9.3